Adding WHIP Server
- Adding WHIP Server binaries

NOTES 1: WHIP Server is the replacement for the asset server.  It was intended to be a more stable replacement to the OpenSim Asset Server.

NOTES 2: While WhiteCoreSim and Virtual Universe both use FileBasedAssetStorage which allow regions to store the assets for content rezzed in-world on the regions on a particular region server for faster rezzing, this is in fct not very stable nor a wise idea.  By storing the assets at the region server level you are in fact creating situations where the region server can become very unstable because it cannot read a asset file, which causes the region server to crash and thus those regions are down until the grid operator restarts that region server.  The Second Galaxy Development Team is working to design a new asset server similar to WHIP to replace the current way Virtual Universe stores assets.

NOTES 3: In additional testing while rezzing may seem faster with FileBasedAssetStorage it is by no means stable.  In fact this adds more unnecessary calls and requirements and CPU usage to the region server in a way that could eventually over time slow down the region services causing timeouts, and serious crashing of the region services.  For grids this is very bad.