Begin moving directories that should be configurable to constants
- Moves the default assets directory to constants
- Moves the default inventory directory to constants
- Moves the default crash logging directory to constants
- Moves the default data directory to constants
- Points the default crash log directory to the directory defined in constants and enables crash logging
- Adds the default logs directory for storage of logs
- Adds the Configuration directory (configuration filees such as the ini files will be moved here soon).
- Copy the assets, data, and inventory directories into the new Data directory folder

NOTE QUESTION: Why am I doing this?

NOTE ANSWER: The reason is very simple.  these are directories that should not really be in the straight bin directory.  This is because when a user copies updated binary files to the bin folder they might accidently overrite a file that will mess up their grid and they will automatically think that theres an error that is the fault of the developer which may just be caused by them accidently overwriting a file they shouldn't have.  This just prevents that possibility.

Both Halcyon and OpenSim do not do this however they should.  Both WhiteCoreSim and the Second Galaxy Development Team's Virtual Universe project arm already acknowledge this and have implemented this.