Moving directories to the Release directory
- Move the bin directory to Release/bin
- Move Configuration directory to Release/Configuration
- Move the Data directory to Release/Data (this moves the sub directories in the Data directory)
- Adds a Licenses directory to the Release/Licenses directory
- Moves BuildingTools to the Release/BuildingTools directory
- Moves the HelperDocs directory to Release/HelperDocs
- Moves the Website directory to Release/Website directory
- Moves the Whip-Server directory to Release/Whip-Server directory (Its already prebuilt version anyway)
- Correct directory urls in Constants.cs
- Update all csproj files
- Update other directory urls where required

NOTE: This will give us the beginning of how the source code directories should be split up from the built release directories.  The Release directory and its contents are what we will provide when we release a new version.  This will help ensure the download size of the Release.tgz or Release.zip will be minimal.  We will also provide a full copy of the source code directory as well i.e. Source.tgz or Source.zip