Map Tile Colors
- Change the map tile colors
- Allow setting of the four terrain textures
- Allow setting of the water color
- Add settings to the Halcyon.ini.Sample file
- Resharp files
- Cleanup commented out code that is either dead or otherwise unused
- Cleanup comments

Note 1: This implements the ability to change the water color per Issue https://github.com/emperorstarfinder/My-Halcyon/issues/20

Note 2: This does not affect Anaximander which handles the map colors on the website world map.  This only affects the color of the water and terrain tiles in the world map the viewer sees.

Note 3: By removing commented out code we reduce the chances that the built binary files will end up with unnecessary bloating which in the long term can potentially slow the architecture down.  The removal of commented out code already gets archived in the git repository so there really is no reason to keep it in the codebase locally itself other then being lazy which really is just irresponsible development.